Admins can no longer create or edit banners. Patron are now responsible for managing banners.
