--- src/main/webapp/META-INF/views/administrator/banner/i18n_en.messages
@@ -1,15 +1,0 @@
-administrator.banner.form.label.picture = Picture
-administrator.banner.form.label.slogan = Slogan
-administrator.banner.form.label.targetURL = URL
-administrator.banner.form.legend.creditCard = Credit card
-administrator.banner.form.button.return = Return
-administrator.banner.form.button.update = Update
-administrator.banner.form.button.delete = Delete
-administrator.banner.form.button.create = Create
-
-administrator.banner.list.label.picture = Picture
-administrator.banner.list.label.slogan = Slogan
-administrator.banner.list.label.targetURL = URL
-
-administrator.banner.error.past-year = The year is in past
-administrator.banner.error.past-month = The month is in past
