--- src/main/webapp/META-INF/views/administrator/banner/i18n_es.messages
@@ -1,16 +1,0 @@
-
-administrator.banner.form.label.picture = Imagen
-administrator.banner.form.label.slogan = Eslogan
-administrator.banner.form.label.url = Dirección
-administrator.banner.form.legend.creditCard = Tarjeta de crédito
-administrator.banner.form.button.return = Volver
-administrator.banner.form.button.update = Actualizar
-administrator.banner.form.button.delete = Eliminar
-administrator.banner.form.button.create = Crear
-
-administrator.banner.list.label.picture = Imagen
-administrator.banner.list.label.slogan = Eslogan
-administrator.banner.list.label.url = Dirección
-
-administrator.banner.error.past-year = El año está en pasado
-administrator.banner.error.past-month = El mes está en pasado
