+++ src/main/webapp/META-INF/views/master/i18n_en.messages
@@ -42,8 +42,6 @@ master.menu.administrator.dashboard = Dashboard
 master.menu.administrator.shutdown = Shut down
 master.menu.administrator.notice.create = Create notice
 master.menu.administrator.notice.list = List of notices
-master.menu.administrator.listBanners = List banners
-master.menu.administrator.createBanners = Create banner
 master.menu.administrator.listInquiries = List inquiries
 master.menu.administrator.createInquiries = Create inquiry
 master.menu.administrator.listOvertures = List overtures
@@ -54,6 +52,13 @@ master.menu.administrator.createTechnologyRecords = Create technology records
 master.menu.administrator.listTechnologyRecords = List technology records
 master.menu.administrator.createToolRecords = Create tool records
 master.menu.administrator.listToolRecords = List tool records
+
+# master.menu.patron ########################################################
+
+master.menu.patron = Patron
+master.menu.patron.listBanners = List banners
+master.menu.patron.createBanners = Create banner
+
 # master.menu.provider ########################################################
 
 master.menu.provider = Provider
