+++ src/main/webapp/META-INF/views/master/i18n_es.messages
@@ -42,8 +42,6 @@ master.menu.administrator.dashboard = Dashboard
 master.menu.administrator.shutdown = Parar el sistema
 master.menu.administrator.notice.list = Lista de noticias
 master.menu.administrator.notice.create = Crear noticia
-master.menu.administrator.listBanners = Listar banners
-master.menu.administrator.createBanners = Crear banner
 master.menu.administrator.listInquiries = Listar peticiones
 master.menu.administrator.createInquiries = Crear petición
 master.menu.administrator.listOvertures = Listar propuestas
@@ -54,6 +52,12 @@ master.menu.administrator.createTechnologyRecords = Crear registros tecnológico
 master.menu.administrator.listTechnologyRecords = Lista de registros tecnológicos
 master.menu.administrator.createToolRecords = Crear registros de herramientas
 master.menu.administrator.listToolRecords = Lista de registros de herramientas
+
+# master.menu.patron ########################################################
+
+master.menu.patron = Patrocinador
+master.menu.patron.listBanners = Listar banners
+master.menu.patron.createBanners = Crear banner
 
 # master.menu.provider ########################################################
