+++ src/main/webapp/META-INF/views/patron/banner/i18n_en.messages
@@ -1,0 +1,15 @@
+patron.banner.form.label.picture = Picture
+patron.banner.form.label.slogan = Slogan
+patron.banner.form.label.targetURL = URL
+patron.banner.form.legend.creditCard = Credit card
+patron.banner.form.button.return = Return
+patron.banner.form.button.update = Update
+patron.banner.form.button.delete = Delete
+patron.banner.form.button.create = Create
+
+patron.banner.list.label.picture = Picture
+patron.banner.list.label.slogan = Slogan
+patron.banner.list.label.targetURL = URL
+
+patron.banner.error.past-year = The year is in past
+patron.banner.error.past-month = The month is in past
