+++ src/main/webapp/META-INF/views/patron/banner/tiles.xml
@@ -18,12 +18,12 @@
 
 <tiles-definitions>
 
-	<definition name="administrator/banner/list" extends="master">
-		<put-attribute name="body" type="template" value="../administrator/banner/list.jsp" />
+	<definition name="patron/banner/list" extends="master">
+		<put-attribute name="body" type="template" value="../patron/banner/list.jsp" />
 	</definition>
 	
-	<definition name="administrator/banner/form" extends="master">
-		<put-attribute name="body" type="template" value="../administrator/banner/form.jsp" />
+	<definition name="patron/banner/form" extends="master">
+		<put-attribute name="body" type="template" value="../patron/banner/form.jsp" />
 	</definition>
 
 </tiles-definitions>
